+++ dev/src/views/ui/attachment.vue
@@ -1,10 +1,10 @@
 <template>
   <div>
     <XAttachment
-      size="default"
+      size="small"
       v-model="fileList"
       v-model:select-value="selected"
-      :limit="2"
+      :limit="5"
       @change="onChange"
       @click="onClick"
       :uploader="uploader"
@@ -13,10 +13,15 @@
       selectable
       :previewable="true"
       :removable="true"
-      :downloadable="false"
-      :multiple="false"></XAttachment>
+      :downloadable="true"
+      :multiple="true">
+      <template #tip>
+        <div>提示文字提示文字提示文字提示文字提示文字</div>
+      </template>
+    </XAttachment>
+    <hr />
     <XAttachment
-      size="default"
+      size="large"
       list-type="list"
       v-model="fileList"
       v-model:select-value="selected"
@@ -32,22 +37,30 @@
       :multiple="true"></XAttachment>
     <!-- <XAttachment v-model="fileList"></XAttachment>
     <XAttachment size="large" v-model="fileList"></XAttachment> -->
+    <hr />
+    <XAttachment
+      multiple
+      :formatter="formatter"
+      :valueFormatter="valueFormatter"
+      :uploader="uploader"
+      v-model="files"
+      :auto-upload="true"></XAttachment>
   </div>
 </template>
 <script lang="ts" setup>
   import { ref } from 'vue';
-  import { XAttachment, request, type AttachmentFile } from '@vtj/web';
+  import { XAttachment, request, type AttachmentFile, delay } from '@vtj/web';
 
   const fileList = ref<AttachmentFile[]>([
     {
       url: 'http://dummyimage.com/120x90',
-      name: 'food.jpeg',
-      type: 'img'
+      name: 'food.jpeg'
+      // type: 'img'
     },
     {
       url: 'http://dummyimage.com/200x300',
-      name: 'food.jpeg',
-      type: 'img'
+      name: 'http://dummyimage.com/200x300food.jpeg'
+      // type: 'img'
     },
     {
       url: 'http://dummyimage.com/300x300/FF0000',
@@ -55,9 +68,9 @@
       type: 'img'
     },
     {
-      url: 'http://dummyimage.com/300x300',
-      name: 'food.docx',
-      type: 'word'
+      url: 'http://dummyimage.com/300x300'
+      // name: 'food.docx'
+      // type: 'word'
     }
   ]);
 
@@ -65,27 +78,27 @@
     url: 'http://dummyimage.com/120x90'
   });
 
-  const uploader = (file: File) => {
-    // return new Promise(() => {});
-    return request({
-      url: '/mock-api/upload',
-      settings: {
-        type: 'data',
-        originResponse: false,
-        validSuccess: true,
-        validate: (r) => r.data.code === 0
-      },
-      data: {
-        file
-      }
-    }).then((res) => {
-      // console.log('request', res);
-      return {
-        name: file.name,
-        url: res as unknown as string
-      };
-    });
-  };
+  // const uploader = (file: File) => {
+  //   // return new Promise(() => {});
+  //   return request({
+  //     url: '/mock-api/upload',
+  //     settings: {
+  //       type: 'data',
+  //       originResponse: false,
+  //       validSuccess: true,
+  //       validate: (r) => r.data.code === 0
+  //     },
+  //     data: {
+  //       file
+  //     }
+  //   }).then((res) => {
+  //     // console.log('request', res);
+  //     return {
+  //       name: file.name,
+  //       url: res as unknown as string
+  //     };
+  //   });
+  // };
 
   const onChange = (files: any) => {
     console.log('onChange', files);
@@ -94,4 +107,37 @@
   const onClick = (file: any) => {
     console.log('click', file);
   };
+
+  const uploader: any = async () => {
+    await delay(1000);
+    // return null;
+    return 'https://oss.newpearl.com/newpearl/image/2024-07-15/acd6ff3e0bf8fce74d795a870c9069e6.png';
+  };
+
+  const formatter: any = async (files: any) => {
+    console.log('formatter', files);
+    for (const file of files) {
+      file.__url = file.url;
+      file.url = file.url + '?only=true';
+    }
+    return files;
+  };
+
+  const valueFormatter: any = async (files: any) => {
+    console.log('valueFormatter', files);
+
+    // for (const file of files) {
+    //   file.url = file.__url;
+    //   delete file.__url;
+    // }
+    // console.log('valueFormatter', files);
+    return files.map((n: any) => {
+      return {
+        url: n.url,
+        name: n.name
+      };
+    });
+  };
+
+  const files = ref([]);
 </script>