+++ packages/designer/src/framework/designer.ts
@@ -214,7 +214,8 @@ export class Designer {
     const { el, model } = helper;
     const refs = el.__context__?.__refs;
     if (!refs) return null;
-    return refs[id || model.id];
+    const instance = refs[id || model.id];
+    return instance?._?.exposed || instance;
   }
 
   private async onDrop(e: DragEvent) {