+++ packages/ui/src/components/qr-code/index.ts
@@ -1,0 +1,3 @@
+import XQrCode from './QrCode.vue';
+export { XQrCode };
+export * from './types';