+++ packages/utils/src/util.ts
@@ -34,3 +34,34 @@ export function formDataToJson(data: FormData) {
   });
   return json;
 }
+
+/**
+ *  Base64 转换 Blob
+ * @param dataurl
+ * @returns
+ */
+export function dataURLtoBlob(dataurl: string) {
+  const arr = dataurl.split(',');
+  const mime = arr[0].match(/:(.*?);/)?.[1];
+  const bstr = atob(arr[1]);
+  let n = bstr.length;
+  const u8arr = new Uint8Array(n);
+  while (n--) {
+    u8arr[n] = bstr.charCodeAt(n);
+  }
+  return new Blob([u8arr], { type: mime });
+}
+
+/**
+ * Blob 转 File
+ * @param blob
+ * @param fileName
+ * @returns
+ */
+export function blobToFile(blob: Blob, fileName: string) {
+  const file = blob as any;
+  file.lastModified = Date.now();
+  file.lastModifiedDate = new Date();
+  file.name = fileName;
+  return file as File;
+}